refactor: job 패키지 생성
